fix: let pdf build success (#1)
This commit enables local PDF build of the documentation. For Mac users, it would be sufficient to install [MacTex](https://tug.org/mactex/) to build the PDF in local machine.

* Use `xelatex` engine for better Unicode support.
* Use three main fonts for local build:
  - 나눔스퀘어(Nanum Square) for Korean
  - Montserrat for English
  - JetBrains Mono for monospaced characters
* Cover design.
* Admonitions are emphasized by colors.

Note that automatic PDF conversion in readthedocs.org does not seem to provide an option to use custom fonts. We may have to use a different `conf.py` to support the remote build, which uses basic fonts available in the readthedocs. That will be handled in another issue.